Coding - MSVC warning fix for STEP Rendering properties #498
Fixed warning on MSVC compiler with name overriding and type mismatching.

+++ src/DataExchange/TKDESTEP/STEPConstruct/STEPConstruct_RenderingProperties.hxx
@@ -144,7 +144,7 @@ public:
   //! @return ColorRGBA object
   Quantity_ColorRGBA GetRGBAColor() const
   {
-    return Quantity_ColorRGBA(mySurfaceColor, 1.0 - myTransparency);
+    return Quantity_ColorRGBA(mySurfaceColor, static_cast<float>(1.0 - myTransparency));
   }
 
   //! Returns surface color
